fix: Unable to generate diff when excluding a helm chart that was previously included
